cleanup application definition
